requirements.txt
influxdb-client
pyairahome
colorlog
python-dotenv
schedule
flask